fix compiler warning
git-svn-id: svn+ssh://svn.abisource.com/svnroot/link-grammar/trunk@33096 bcba8976-2d24-0410-9c9c-aab3bd5fdfd6

+++ link-parser/command-line.c
@@ -354,7 +354,7 @@ static int x_issue_special_command(const char * line, Parse_Options opts, Dictio
 		else
 		{
 			float val = -1.0;
-			val = atof(y);
+			val = (float) atof(y);
 			if (val < 0.0f)
 			{
 				printf("Invalid value %s for variable %s Type \"!help\" or \"!variables\"\n", y, as[j].string);
